# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "Request an OTP to the given phone number or email.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully verified, returns tokens."
          },
          "401": {
            "description": "Invalid or expired OTP."
          }
        },
        "summary": "Verify the received OTP code.",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully refreshed tokens."
          },
          "401": {
            "description": "Invalid refresh token."
          }
        },
        "summary": "Rotate the refresh token for new access/refresh tokens.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/matches/{matchId}/devices/public-keys": {
      "get": {
        "operationId": "MessagesController_getMatchPublicKeys",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get public keys of the matched user for E2E setup",
        "tags": [
          "messages"
        ]
      }
    },
    "/matches/{matchId}/messages": {
      "post": {
        "operationId": "MessagesController_createMessage",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a new encrypted message",
        "tags": [
          "messages"
        ]
      },
      "get": {
        "operationId": "MessagesController_getMessages",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get paginated messages for a match",
        "tags": [
          "messages"
        ]
      }
    },
    "/matches/{matchId}/messages/{messageId}/read": {
      "patch": {
        "operationId": "MessagesController_markAsRead",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a message as read",
        "tags": [
          "messages"
        ]
      }
    },
    "/matches/{matchId}/messages/attachments/upload-url": {
      "post": {
        "operationId": "MessagesController_getAttachmentUploadUrl",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get signed URL to upload an attachment",
        "tags": [
          "messages"
        ]
      }
    },
    "/matches/{matchId}/messages/{messageId}/attachments": {
      "post": {
        "operationId": "MessagesController_registerAttachment",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register an uploaded attachment",
        "tags": [
          "messages"
        ]
      }
    },
    "/matches/{matchId}/messages/{messageId}/attachments/{attachmentId}/reveal": {
      "post": {
        "operationId": "MessagesController_revealAttachment",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reveal an attachment",
        "tags": [
          "messages"
        ]
      },
      "delete": {
        "operationId": "MessagesController_revokeAttachment",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke an attachment reveal",
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lib le Lib API",
    "description": "API documentation for the Lib le Lib application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string",
            "description": "Phone number or email address to send the OTP to.",
            "example": "+251911000000"
          },
          "isSignUp": {
            "type": "boolean",
            "description": "True if this is a signup attempt, false if login.",
            "example": true
          }
        },
        "required": [
          "destination",
          "isSignUp"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "destination": {
            "type": "string",
            "description": "Phone number or email address the OTP was sent to.",
            "example": "+251911000000"
          },
          "code": {
            "type": "string",
            "description": "The 6-digit OTP code.",
            "example": "123456"
          },
          "isSignUp": {
            "type": "boolean",
            "description": "True if this is a signup attempt, false if login.",
            "example": true
          }
        },
        "required": [
          "destination",
          "code",
          "isSignUp"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token issued at login."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "messageType": {
            "type": "string",
            "enum": [
              "text",
              "image"
            ],
            "default": "text"
          },
          "ciphertext": {
            "type": "string",
            "description": "Base64 encoded ciphertext"
          },
          "nonce": {
            "type": "string",
            "description": "Base64 encoded nonce"
          }
        },
        "required": [
          "messageType",
          "ciphertext",
          "nonce"
        ]
      },
...
    }
  }
}